Server - Fix races and deadlocks with backend updates support.
Support for backend updates had a lot of data races and could deadlock.

The client's subscription (username, mailbox name, silent flag) need to
be protected by the server lock.

The update from the backend should be written to a buffer before sending
to the clients, because some responses used channels that couldn't be
read from multiple times (only some clients would get a partial response).

A client connection might already be handling a command response, which
would block.  I think it is best to just skip those clients for now.